refactor(ui): simplify chain-of-thought runtime and trace architecture
