+++ packages/ui/src/components/assistant-ui/chain-of-thought/runtime.tsx
@@ -1,5 +1,5 @@
 "use client";
 
-export { ChainOfThoughtImpl, type ChainOfThoughtProps } from "./runtime-impl";
+export { ChainOfThoughtImpl, type ChainOfThoughtProps } from "./runtime-root";
 export type { ToolActivity } from "./runtime-activity";
 export { ChainOfThoughtTraceTool } from "./runtime-tool";
